move patchseq_history to aisynphys and rework
no longer in notes_db, meant to be run in a notebook to get data when interested
